#110 refactored component tests
