+++ storage/src/test/java/org/zalando/apidiscovery/storage/ApiEntityToApiDefinitionConverterTest.java
@@ -6,16 +6,17 @@ import org.zalando.apidiscovery.storage.api.ApiDeploymentEntity;
 import org.zalando.apidiscovery.storage.api.ApiEntity;
 import org.zalando.apidiscovery.storage.api.ApplicationEntity;
 import org.zalando.apidiscovery.storage.api.DeploymentLinkDto;
+import org.zalando.apidiscovery.storage.utils.DomainObjectGen;
 
 import static java.util.Arrays.asList;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.API_UI;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.API_URL;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.LIFECYCLE_STATE;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.NOW;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.givenApiDeployment;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.givenApiEntity;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.givenApplication;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.API_UI;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.API_URL;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.LIFECYCLE_STATE;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.NOW;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.givenApiDeployment;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.givenApiEntity;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.givenApplication;
 import static org.zalando.apidiscovery.storage.api.ApiEntityToApiDefinitionConverter.toApiDefinitionDto;
 
 public class ApiEntityToApiDefinitionConverterTest {
