+++ storage/src/test/java/org/zalando/apidiscovery/storage/ApiVersionResourceIntegrationTest.java
@@ -14,17 +14,17 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasItem;
 import static org.hamcrest.Matchers.hasItems;
 import static org.hamcrest.Matchers.hasSize;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.API_NAME;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.API_UI;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.API_URL;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.API_VERSION_1;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.API_VERSION_2;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.APP1_NAME;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.DEFINITION;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.DEFINITION_TYPE;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.givenApiDeployment;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.givenApiEntity;
-import static org.zalando.apidiscovery.storage.DomainObjectGen.givenApplication;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.API_NAME;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.API_UI;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.API_URL;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.API_VERSION_1;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.API_VERSION_2;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.APP1_NAME;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.DEFINITION;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.DEFINITION_TYPE;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.givenApiDeployment;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.givenApiEntity;
+import static org.zalando.apidiscovery.storage.utils.DomainObjectGen.givenApplication;
 import static org.zalando.apidiscovery.storage.api.ApiLifecycleState.ACTIVE;
 import static org.zalando.apidiscovery.storage.api.ApiLifecycleState.DECOMMISSIONED;
 import static org.zalando.apidiscovery.storage.api.ApiLifecycleState.INACTIVE;
