--- storage/src/test/java/org/zalando/apidiscovery/storage/ResourceComponentTestConfig.java
@@ -1,12 +1,0 @@
-package org.zalando.apidiscovery.storage;
-
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-
-@Configuration
-@Import({
-    JacksonConfiguration.class
-})
-public class ResourceComponentTestConfig {
-}
